requirements.txt
Django==1.6.7
Fabric==1.10.0
MySQL-python==1.2.5
PyYAML==3.11
South==1.0
Werkzeug==0.9.6
argparse==1.2.1
boto==2.32.1
dj-database-url==0.3.0
dj-static==0.0.6
django-bootstrap-toolkit==2.15.0
django-debug-toolbar==1.2.1
django-extensions==1.4.0
django-herokuapp==0.9.18
django-registration==1.0
django-require==1.0.6
django-require-s3==1.0.0
django-storages==1.1.8
django-suit==0.2.11
django-toolbelt==0.0.1
docutils==0.12
ecdsa==0.11
gunicorn==19.1.1
paramiko==1.15.1
psycopg2==2.5.4
py-mysql2pgsql==0.1.5
pycrypto==2.6.1
pytz==2014.7
sh==1.09
six==1.7.3
sqlparse==0.1.11
static3==0.5.1
termcolor==1.1.0
waitress==0.8.9
wsgiref==0.1.2
